refactor: ProductItem 커스텀 훅 분리
